refactor: upgrade order ID to distributed Snowflake ID
- Refactor Order entity to use orderId (Long) as primary key
- Update OrderRepository to use Long as ID type
- Update PaymentResponse DTO to use Long for orderId
- Add migration script for existing databases
- Update schema-with-userid.sql to include orderId
- Add comprehensive ORDERID_REFACTORING_GUIDE.md documentation

Benefits:
- Prevents business data leakage (order volume, revenue)
- Supports order sharding and distributed deployment
- Improves payment callback security and reliability
- Enables cross-system order reconciliation
- Maintains backward compatibility with physical id field

+++ server-java/src/main/java/com/usjobhunt/repository/OrderRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     Optional<Order> findByTransactionId(String transactionId);
-    List<Order> findByUserId(Integer userId);
+    List<Order> findByUserId(Long userId);
 }
